move web specific dropdown utils to web file
